tests/wpt/mozilla/tests/mozilla/sequence-hole.html

<!doctype html>
<meta charset="utf-8">
<title></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  var t = new TestBinding;
  assert_throws(new TypeError, function () { t.passInterfaceSequence([new Blob,, new Blob]) });
}, "Sequences with holes are correctly handled");
</script>
